backend/requirements.txt
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.32
langchain-chroma==0.1.0
langchain-text-splitters==0.0.1
psycopg2-binary
sqlalchemy
pgvector
langchain-postgres==0.0.1
faiss-cpu
chromadb
fastapi
uvicorn[standard]
python-multipart
beautifulsoup4
requests
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
pypdf
tiktoken
openai
